Initial macOS build
Got XPNet.Native, XPNet.CLR and XPNet.LoggerPlugin building and working on macOS, with some caveats: it only does a 64-bit build (no 32-bit build yet), and currently needs boost at build time.  The boost dependency is not intended in the long term, but it's needed until the default macOS build tools support C++17 filesystem).  This does not get the test harness running on macOS; I don't really need it to, so I'll leave that task for someone who does need or want it to.