Fixes for automatic runtime installation and selection.
Corrected automatic dotnet runtime install locations in the template project: always install into a arch-specific subdirectory now.

Correctly identify the plugin directory for the case where the dotnet runtime is detect in an arch-specific subdirectory.

In our native XPluginStart callback, if any of the dotnet plugin methods cannot be resolved (like ClrPluginStart, for example because the XPNet.CLR.dll isn't there or can't be loaded), log a message and bail out instead of crashing.

Before these changes, detection happened to work for the cases I was testing but would in other cases choose a mismatching runtime (trying to load a 64-bit dotnet runtime in a 32-bit process, for instance).