Automating the Windows build, with a few macOS build improvements.
Misc. improvements to the Unix Makefile; added test targets to it and made the package target depend on the tests passing.

Added Makefile.vc, so that nmake can be used to build on Windows like make on Unix.  It supports the same top-level targets and has the same general behavior as the Unix Makefile.

Moved the .csproj and .nuspec files for the native packages into their own directory, after tracking down mysterious transient build errors to those projects sharing an 'obj' folder with the XPNet.Native .vcxproj.