Converted from manually-verified tests to automated unit tests.
Added automated unit tests covering most of the things that the manual tests covered, plus a few extra things.  A new XPNet.CLR.TestPlugins project contains the CLR side of those tests, as a set of tiny test-specific IXPlanePlugin classes (in files like FlightLoopTestPlugin.cs).  Those plugins are paired with C++ tests written in Google Test (in files like FlightLoopTests.cpp, etc.)

Added XPBoolean: this fixes a bug found while creating the new unit tests, where the bool[] DataRef mapping was incorrectly mapping 8-bit booleans on the .NET side to 32-bit integers on the native side.  The crux: a single 8-bit boolean _is_ guaranteed to take up 32-bits on the stack, but an _array_ of 8-bit booleans are packed together with no extra padding, so DataRef<bool> worked but DataRef<bool[]> did not.  XPBoolean is a struct containing an int with operators/methods to make it act like a bool, so you can still work with bool-like semantics on the .NET side even though X-Plane uses int to represent boolean values.  The tests included in this checkin _do_ include a regression test for this mapping.

Changes to the test harness to support the new unit tests: 'get' functions to read back datarefs written by the test plugins; 'clear' functions to clear out setup between tests so that datarefs/commands/etc. set up in one test do not affect other tests; ability to set a plugin path different than the path of the executable running the tests; removed the console writes (via 'cout') that had been added to the harness, since they support manual test verification and the intent of this harness is to support automated test verification.  NOTE: _Some_ of the console writes do not have a replacement as of this checkin, particularly in the graphics tests; we'll need to come behind and add additional unit tests to cover the graphics functionality that isn't yet covered here.

Fixed an unhandled exception trying to log to a null logger in XPNetPluginBridge in edge error cases; the case couldn't really happen if the host environment never calls the plugin entry points in an invalid order, but that occasionally happened in unit test development and the extra protection doesn't hurt anything.  Related to the next entry.

Added instrumentation to XPNetPlugin.cpp for the test environment to query the state of the plugin (isStarted, isEnabled) to avoid making invalid calls to it during test teardown (e.g., after tests fail, or a failure test succeeds).  This could have been done purely at the testing layer, but it's much more straighforward and less likely to break in maintenance if added to the plugin itself.

Removed all the extraneous code from LoggerPlugin that was only there to support the older test cases.  LoggerPlugin is not used in the test cases anymore, and is _just_ a sample plugin now.

Require C# language 7.2 support to compile XPNet.CLR, allowing us to use newer stable language features.

Add new TESTING.md describing how the tests are set up, and Updated README.md for the new Windows command-line build and test setup.